configuration des autorisations aux endpoints nécessaires

+++ src/EnchereService/Controllers/EncheresController.cs
@@ -5,6 +5,7 @@ using EnchereService.Data;
 using EnchereService.DTOs;
 using EnchereService.Entities;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -50,12 +51,13 @@ public class EncheresController : ControllerBase
         return _mapper.Map<EnchereDto>(enchere);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<EnchereDto>> CreateEnchere(CreateEnchereDto enchereDto)
     {
         var enchere = _mapper.Map<Enchere>(enchereDto);
-        // TODO : add current user as seller
-        enchere.Seller = "test";
+
+        enchere.Seller = User.Identity.Name;
 
         _context.Encheres.Add(enchere);
 
@@ -71,6 +73,7 @@ public class EncheresController : ControllerBase
         new { enchere.Id }, _mapper.Map<EnchereDto>(enchere));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateEnchere(Guid id, UpdateEnchereDto updateEnchereDto)
     {
@@ -78,7 +81,7 @@ public class EncheresController : ControllerBase
         .FirstOrDefaultAsync(x => x.Id == id);
         if (enchere == null) return NotFound();
 
-        //TODO : check seller == username
+        if (enchere.Seller != User.Identity.Name) return Forbid();
 
         enchere.Produit.Make = updateEnchereDto.Make ?? enchere.Produit.Make;
         enchere.Produit.ProductName = updateEnchereDto.ProductName ?? enchere.Produit.ProductName;
@@ -96,6 +99,7 @@ public class EncheresController : ControllerBase
         return BadRequest("Erreur lors de la sauvegarde");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteEnchere(Guid id)
     {
@@ -103,7 +107,7 @@ public class EncheresController : ControllerBase
 
         if (enchere == null) return NotFound();
 
-        //TODO : check seller == username
+        if (enchere.Seller != User.Identity.Name) return Forbid();
 
         _context.Encheres.Remove(enchere);
